requirements.txt
# Core Web Framework
flask>=2.0.0
werkzeug>=2.0.0

# Data Analysis & Processing
pandas>=1.3.0
numpy>=1.21.0

# Financial Data APIs
yfinance>=0.2.0

# Social Media APIs
tweepy>=4.0.0

# AI & Machine Learning
anthropic>=0.7.0
openai>=1.0.0

# Web Scraping
beautifulsoup4>=4.10.0
selenium>=4.0.0
requests>=2.25.0

# Data Visualization
plotly>=5.0.0
matplotlib>=3.3.0

# Database
sqlite3

# Utilities
python-dotenv>=0.19.0
schedule>=1.1.0
pytz>=2021.3

# Optional dependencies for enhanced features
praw>=7.0.0  # Reddit API (optional)
newsapi-python>=0.2.6  # News API (optional)

# Web framework
flask-cors==4.0.0
jinja2==3.1.2

# Email
sendgrid==6.10.0

# Data processing
python-dateutil==2.8.2

# Visualization
seaborn==0.13.0
fake-useragent==1.4.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
schedule==1.2.0

# Social Media APIs
praw==7.7.1
tweepy==4.14.0

# Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# News API
newsapi-python==0.2.7

# Async processing
aiohttp==3.9.1

# Additional ML packages
scikit-learn==1.3.0

# Database
pymongo==4.6.1
redis==5.0.1

# Feed parsing
feedparser==6.0.10 